+++ Python/NeonOcean.S4.Refer/NeonOcean/S4/Refer/Settings/CustomPronounSets.py
@@ -1,7 +1,0 @@
-from NeonOcean.S4.Refer.Settings import Types as SettingsTypes
-
-class CustomPronounSets(SettingsTypes.CustomPronounSetsDialogSetting):
-	IsSetting = True  # type: bool
-
-	Key = "Custom_Pronoun_Sets"  # type: str
-	Default = dict()  # type: dict