+++ Python/NeonOcean.S4.Refer/NeonOcean/S4/Refer/Settings/Types.py
@@ -27,3 +27,4 @@ class CustomPronounSetsSetting(SettingsBase.Setting):
 
 class CustomPronounSetsDialogSetting(CustomPronounSetsSetting):
 	Dialog = SettingsDialogs.CustomPronounSetsDialog
+	EditPronounSetDialog = SettingsDialogs.EditPronounSetDialog