+++ Python/NeonOcean.S4.Refer/NeonOcean/S4/Refer/ThisNamespace.py
@@ -1,0 +1,1 @@
+Namespace = "NeonOcean.S4.Refer"  # type: str